WIP: BLSLA-13: add library to disable cmp if needed
We keep this library implementation as backup. It is the correct implementation of this function and it should work and it does in most cases. But sometimes the __tcfapi object/function is ready to late and the toggle of the consent manager does not work.
We implemented this function via inline script for Superillu and Guterrat